patch1 changes
Graph by django-filter
choices updated
jsonresponse dictionary data
dataupdate with serializer
anonymous user for authentication

+++ price/models.py
@@ -3,7 +3,6 @@ from company.models import Company
 
 
 class Price(models.Model):
-
     company = models.ForeignKey(Company, on_delete=models.CASCADE, related_name='company_price')
     ldcp = models.FloatField()
     open = models.FloatField()
@@ -11,5 +10,6 @@ class Price(models.Model):
     low = models.FloatField()
     current = models.FloatField()
     change = models.FloatField()
-    volume = models.FloatField()
+    volume = models.IntegerField()
     date_time = models.DateTimeField()
+
